Resolved the statement number to symbol type problem a new way.
I never liked the cast.  I didn't want to define statement_number_t in
symbols.  I didn't want to intentionally use unsigned long.  So the
symbol name constructor will now accept any unsigned integral type.

+++ fortran/basicstatement.cpp
@@ -9,10 +9,9 @@ namespace aid::fortran {
 std::string basic_statement::generate(unit const &u) const {
     auto const code = do_generate(u);
     if (m_number == no_statement_number) return code;
-    auto const symbol = u.find_symbol(symbol_name{static_cast<unsigned int>(m_number)});
+    auto const symbol = u.find_symbol(symbol_name{m_number});
     if (!symbol.referenced) return code;
     return std::format("L{0}: {1}", m_number, code);
-    //return std::format("L{0}: puts(\"L{0}:\"); {1}", m_number, code);
 }
 
 void basic_statement::mark_reachable(unit &u) {
